Move GritTransformerLayer to components/grit_layers.py
